create migrations and connect models in telegram bot, some litle change in tables and add migrations

+++ mini_kinopoisk/films_serials/models.py
@@ -8,7 +8,7 @@ class SearchFilm(models.Model):
     country = models.CharField(null=True)
     movie_length = models.PositiveIntegerField(null=True)
     description = models.TextField(null=True)
-    rating = models.DecimalField(max_digits=1, decimal_places=1, null=True)
+    rating = models.FloatField(null=True)
     age_rating = models. PositiveBigIntegerField(null=True)
     picture = models.URLField(max_length=200, null=True)
 
@@ -16,7 +16,7 @@ class SearchFilm(models.Model):
 class SearchSerial(models.Model):
     name = models.CharField(max_length=60, null=True)
     janr = models.CharField(max_length=60, null=True)
-    rating = models.DecimalField(max_digits=1, decimal_places=1, null=True)
+    rating = models.FloatField(null=True)
     release_year = models.CharField(max_length=60, null=True)
     series_length = models.CharField(max_length=10, null=True)
     country = models.CharField(null=True)
